1. 修改员工约谈详情修改“总评”时候未上送的bug 2. 员工主页增加“约谈总时长”的显示字段 3. 修改日程中，无事件也会显示数量的bug 4. 成绩录入中，删去显示custom字段，同时，SAT2和AP这两种自定义科目的输入判断修正。

+++ VisionaryEdu/VisionaryEdu/Model/documents/HonorActivityModel.m
@@ -82,7 +82,7 @@
         }
     }
     
-    NSLog(@"活动数组个数 = %lu, 荣誉数组个数 = %lu",ActivityArray.count, HonorArray.count);
+    NSLog(@"活动数组个数 = %lu, 荣誉数组个数 = %lu",(unsigned long)ActivityArray.count, (unsigned long)HonorArray.count);
     block(HonorArray,ActivityArray);
 }
